use a left outer join for experience types

+++ openedx/core/djangoapps/schedules/migrations/0006_scheduleexperience.py
@@ -15,7 +15,7 @@ class Migration(migrations.Migration):
             name='ScheduleExperience',
             fields=[
                 ('id', models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True)),
-                ('experience_type', models.IntegerField(default=0, choices=[(0, b'Recurring Nudge and Upgrade Reminder'), (1, b'Course Updates')])),
+                ('experience_type', models.PositiveSmallIntegerField(default=0, choices=[(0, b'Recurring Nudge and Upgrade Reminder'), (1, b'Course Updates')])),
                 ('schedule', models.OneToOneField(related_name='experience', to='schedules.Schedule')),
             ],
         ),
